% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport-traffic.R
\name{airport_daily_stats}
\alias{airport_daily_stats}
\title{Title}
\usage{
airport_daily_stats(.apdf, .apt = NULL, .yr = NULL, ...)
}
\arguments{
\item{.apdf}{airport movement table}

\item{.apt}{(optional) trim results to specific airport}

\item{.yr}{(optional) filter for annual results}

\item{...}{potential variables}
}
\value{
tibble of traffic counts
}
\description{
Title
}
\examples{
\dontrun{
  airport_daily_stats(my_apdf_data, "LSZH")
}
}
