% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansperf-package.R
\docType{package}
\name{ansperf-package}
\alias{ansperf}
\alias{ansperf-package}
\title{ansperf: Operational Performance Benchmarking with R}
\description{
Utility functions in support of the bi-regional and multi-regional benchmarking. Implements the ICAO GANP KPIs and associated PBWG or bi-lateral variants.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://euctrl-pru.github.io/ansperf/}
}

}
\author{
\strong{Maintainer}: Rainer Koelle \email{rainer.koelle@eurocontrol.int} (\href{https://orcid.org/0000-0003-2810-1404}{ORCID})

}
\keyword{internal}
