% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ansperf.R
\docType{package}
\name{ansperf}
\alias{ansperf}
\alias{ansperf-package}
\title{ansperf: a package for the analysis of ANS Performance}
\description{
The ansperf package provides tools to support the analysis
of the performance of air navigation services (ANS).
The aim is to provide a common set of functions for the analysis
and visualisation based on a standard set of performance data.
}
\section{ansperf functions}{

to be developed a set / summary of functions
}

