% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xhelp_append_DOF.R
\name{append_dof}
\alias{append_dof}
\title{Add date-of-flight (DOF) to APDF}
\usage{
append_dof(.flts)
}
\arguments{
\item{.flts}{}
}
\value{
tibble with appended DOF column
}
\description{
Add date-of-flight (DOF) to APDF
}
\examples{
\dontrun{
df_with_DOF <- df |> append_dof
}
}
