% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apt-traffic.R
\name{apt_traffic_count}
\alias{apt_traffic_count}
\title{demo - simple traffic counts of an airport}
\usage{
apt_traffic_count(apdf)
}
\arguments{
\item{apdf}{input airport level data frame/tibble}
}
\value{
tibble of counts
}
\description{
demo - simple traffic counts of an airport
}
\examples{
\dontrun{
  apt_traffic_count(lszh)
}
}
