% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxi-times-references.R
\name{build_txxt_reference}
\alias{build_txxt_reference}
\title{Establish the reference data set for the additional taxi-in/taxi-out method.}
\usage{
build_txxt_reference(refdf, variant = NULL, ref_period = NULL)
}
\arguments{
\item{refdf}{data frame/tibble of taxi-times serving as reference data set}

\item{variant}{of the determination of the reference times (i.e., GANP, PBWG)}

\item{ref_period}{the reference period (currently: year)}
}
\description{
Establish the reference data set for the additional taxi-in/taxi-out method.
}
