% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xhelp-check-vars.R
\name{checkvars}
\alias{checkvars}
\alias{merge_var_mappings}
\alias{check_required_columns}
\alias{check_column_types}
\alias{validate_input_vars}
\title{Helper functions to check and validate user input}
\usage{
merge_var_mappings(defaults, user_args = list())

check_required_columns(data, mapping)

check_column_types(data, mapping, types)

validate_input_vars(data, defaults, user_args = list(), types = NULL)
}
\arguments{
\item{defaults}{\itemize{
\item the expected default variable names for the calculation of the metric
}}

\item{user_args}{\itemize{
\item the user defined argument names
}}

\item{data}{\itemize{
\item input data frame / tibble
}}

\item{mapping}{\itemize{
\item the detected mapping between defaults and user names
}}

\item{types}{\itemize{
\item the required types
}}
}
\description{
A set of functions that checks and validate user inputs.
}
