% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coerce_timestamp.R
\name{coerce_timestamp}
\alias{coerce_timestamp}
\title{ansperf time coercion functions}
\usage{
coerce_timestamp(vec, tz = "UTC")
}
\value{
the coerced vector or time columns of the data frame.
}
\description{
ansperf works on date-timestamps for the flight movement data.
A set of coercion utility functions allow for convenient data
transformation operations.
}
\examples{
coerce_timestamp(vec=timevector, tz="UTC")
}

