% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdata.R
\docType{data}
\name{colspec}
\alias{colspec}
\title{GANP colspec from APDF --------------------------------------------------}
\format{
An object of class \code{col_spec} of length 3.
}
\usage{
colspec
}
\description{
GANP colspec from APDF --------------------------------------------------
}
\keyword{datasets}
