% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ganp-kpi-reqs.R
\docType{data}
\name{data_ganp_kpi_reqs}
\alias{data_ganp_kpi_reqs}
\title{GANP KPI table}
\format{
\subsection{\code{data_ganp_kpi_reqs}}{

A data frame with 22 KPIs and 6 columns, 2 header rows
\describe{
\item{KPI}{indicator and number, e.g. KPI08}
\item{name}{descriptive name of indicator}
\item{data-basic}{data items typically found in a basic flight table dataset}
\item{data-advanced}{either ATFM data, meta-data, or trajectory data required to calculate or derive the trajectory based indicators}
}
}
}
\source{
\url{https://www4.icao.int/ganpportal/ASBU/KPI}
}
\usage{
data_ganp_kpi_reqs
}
\description{
A table listing the current GANP KPIs and associated data items.
}
\keyword{datasets}
