% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apt_info.R
\name{get_apt_info}
\alias{get_apt_info}
\title{Retrieve airport information from airport-data.com via API}
\usage{
get_apt_info(.icao)
}
\arguments{
\item{.icao}{ICAO location indicator}
}
\value{
tibble of airport information
}
\description{
Retrieve airport information from airport-data.com via API
}
\examples{
\dontrun{
egll <- get_apt_info("EGLL")
egll
}
}
