% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelloWorld.R
\name{hello_world}
\alias{hello_world}
\title{hello_world returns greeting with name}
\usage{
hello_world(.name)
}
\arguments{
\item{.name}{}
}
\value{
greeting phrase with name
}
\description{
hello_world returns greeting with name
}
\examples{
hello_world("Rainer")
}
