% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xdata.R
\name{load_apdf}
\alias{load_apdf}
\title{Load example APDF data for an airport}
\usage{
load_apdf(.apt, .trim = TRUE)
}
\arguments{
\item{.apt}{character of airport location indicator, either EGLL, LSZH, or LTFM}

\item{.trim}{boolean to remove superflous columns from supplied tibble}
}
\value{
tibble of APDF data set
}
\description{
Load example APDF data for an airport
}
\examples{
\dontrun{
ds <- load_apdf("LSZH")
}
}
