% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-load-apdf.R
\name{make_nice_names_apdf}
\alias{make_nice_names_apdf}
\title{nice names}
\usage{
make_nice_names_apdf(.df)
}
\arguments{
\item{.df}{APDF tibble/dataframe}
}
\description{
utility function to recode/rename column names
}
