% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/session_log.R
\name{session_log}
\alias{session_log}
\title{Information on session}
\usage{
session_log()
}
\arguments{
\item{NULL,}{session_log requires no parameters}
}
\value{
NULL, session_log returns no values
}
\description{
session_log() provides a shortcut to relevant session info
it is a generic function
}
\examples{
session_log()
}

