% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-hello-world.R
\name{x_hello_world}
\alias{x_hello_world}
\title{A cool salutation of a guest}
\usage{
x_hello_world(greeted_actor)
}
\arguments{
\item{greeted_actor}{the person to greet}
}
\value{
salute phrase
}
\description{
A cool salutation of a guest
}
