% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-load-apdf.R
\docType{data}
\name{load_apdf}
\alias{load_apdf}
\alias{read_apdf}
\alias{make_nice_names_apdf}
\alias{colspec}
\alias{trim_ganp_cols}
\title{Load example APDF data for an airport based on European data extraction.}
\format{
An object of class \code{col_spec} of length 3.
}
\usage{
load_apdf(.apt, .trim = TRUE)

read_apdf(.apt)

make_nice_names_apdf(.df)

colspec

trim_ganp_cols(.ds)
}
\arguments{
\item{.apt}{ICAO location indicator of airport}

\item{.trim}{boolean to remove superflous columns from supplied tibble}

\item{.df}{APDF tibble/dataframe}

\item{.ds}{tibble with APDF variables}
}
\value{
tibble of APDF data set
}
\description{
This data is provided for testing purposes.
We need to see whether to provide this data as part of the package to the end user.

utility function to recode/rename column names

define col naming convention for APDF data
}
\examples{
\dontrun{
ds <- load_apdf("LSZH")
}
}
\keyword{datasets}
