% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansperf-package.R
\docType{package}
\name{ansperf-package}
\alias{ansperf}
\alias{ansperf-package}
\title{ansperf: Operational ANS Performance with R}
\description{
Tools to conduct an operational ANS performance analysis with R making use of the indicators developed under ICAO's GANP or used by PBWG and multiple bi-regional projects, e.g. Brazil-Europe.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://euctrl-pru.github.io/ansperf/}
}

}
\author{
\strong{Maintainer}: Rainer Koelle \email{rainer.koelle@eurocontrol.int} (\href{https://orcid.org/0000-0003-2810-1404}{ORCID})

}
\keyword{internal}
