% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punctuality.R
\name{calc_punctuality}
\alias{calc_punctuality}
\alias{check_punc_vars}
\alias{add_delay_and_dlygrp}
\title{Calculate arrival or departure punctuality}
\usage{
calc_punctuality(.apdf)

check_punc_vars(.apdf)

add_delay_and_dlygrp(.apdf)
}
\arguments{
\item{.apdf}{}
}
\value{
tibble of results
}
\description{
What this set of functions should do:
\itemize{
\item check that we have all the required data columns
\item add the delay and delay group
\item calculate the punctuality per phase
}
}
\examples{
\dontrun{
mickey documentation needed
}
}
