% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punctuality.R
\name{calc_punctuality}
\alias{calc_punctuality}
\alias{check_punc_vars}
\alias{add_delay_and_dlygrp}
\alias{pivot_daily_dly_by_grp}
\alias{calc_earlylatewithin_bins}
\title{Calculate arrival or departure punctuality}
\usage{
calc_punctuality(.apdf, .apt)

check_punc_vars(.apdf)

add_delay_and_dlygrp(.apdf)

pivot_daily_dly_by_grp(.dlys, .dly_order = dly_order)

calc_earlylatewithin_bins(.dlys_wide)
}
\arguments{
\item{.apdf}{APDF tibble}

\item{.dlys}{tible of delays}

\item{.dly_order}{vector of sorted ordering}

\item{.dlys_wide}{summary pivot tibble with delays in wide format}
}
\value{
tibble of results

apdf_version vector of boolean flags for format version
}
\description{
What this set of functions should do:
\itemize{
\item check that we have all the required data columns
\item add the delay and delay group
\item calculate the punctuality per phase
}

Early (-15,-5]","Early (-5,0]","Late (0,5)","Late [5,15)","Within (-5,5)","Within (-15,15)"
}
\examples{
\dontrun{
mickey documentation needed
}
}
