% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-load-apdf.R
\docType{data}
\name{colspec}
\alias{colspec}
\title{GANP colspec from APDF}
\format{
An object of class \code{col_spec} of length 3.
}
\usage{
colspec
}
\description{
define col naming convention for APDF data
}
\keyword{datasets}
