% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Hello world - example function}
\usage{
hello_world(.name)
}
\arguments{
\item{.name}{the name used to greet}
}
\value{
a greeting string
}
\description{
Hello world - example function
}
\examples{
hello_world("Rainer")

}
