% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-load-apdf.R
\name{load_apdf}
\alias{load_apdf}
\title{Load example APDF data for an airport based on European data extraction.}
\usage{
load_apdf(.apt, .trim = TRUE)
}
\arguments{
\item{.apt}{character of airport location indicator, either EGLL, LSZH, or LTFM}

\item{.trim}{boolean to remove superflous columns from supplied tibble}
}
\value{
tibble of APDF data set
}
\description{
This data is provided for testing purposes.
We need to see whether to provide this data as part of the package to the end user.
}
\examples{
\dontrun{
ds <- load_apdf("LSZH")
}
}
