% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-hello-world.R
\name{x_hello_world}
\alias{x_hello_world}
\title{Lauding a person}
\usage{
x_hello_world(greeted_actor)
}
\arguments{
\item{greeted_actor}{the person's name to be greeted}
}
\value{
the salutation
}
\description{
Saluting a to-be greated person.
}
\examples{
x_hello_world("Master")

}
